Make a functional game
The road generator generates random positions for the next road segment.
Generator speeds up as more road is generated/passed.
The car explodes if it falls off the road. It can be controlled with a
keyboard. The car owns the main thread and calls the generator as more
road is needed.
A destroyer thread monitors the cars falling off the road (heavy work).
It's not designed to be dynamically rescalable but initialization size
must be set.
Game controls such as quit, pause and return are implemented.