Do a lot of AI stuff
Create a gene that contains chromosomes that contain a sensor&action.
Gene, action and sensor mutations need to be created. Mabye it's better
to leave mutations only to the gene.
Combining genes has a problem, can't figure how to variate number of
chromosomes; needed because the less chromosomes needed the better.
Evaluating should consider the score, number of chromosomes and
strengths(max distance) of sensors.
Number of actions should be considered, but can't because the road is
randomly generated. Should therefore build a static road(course) for bot
testing.