Start building the controller for population, comment the idea, connect and manage threads.
Population controller (as it should be called) now has a thread that,
when started, creates a random population and connects the genome with
the game instance. Each game is currently in it's separate window (hope
so).
Btw Gene is a stupid name so I'm using Genome now (Genome has
chromosomes, which have DNAs which have genes, so...).
Created a method for accessing the road 'bitmap' from outside
RoadGenerator for a point (needed for the poor sensors).
Each Unit has its own thread that will trigger sensors through the
Genome.
Action and Movement are joined in MovementAction, still needs
implementation of mutate() and other stuff.